fix compile warnings

+++ thread2thread2.cc
@@ -38,7 +38,7 @@ static void default_t2t2_assert_handler(t2t2_error_t e,
 {
     fprintf(stderr,
             "\n\nERROR: Thread2Thread2 ASSERTION %d (%s) at %s:%d\n\n",
-            e, t2t2_error_types[(int)e], filename, lineno);
+            (int)e, t2t2_error_types[(int)e], filename, lineno);
     if (fatal)
         // if you dont like this exiting, CHANGE IT
         exit(1);
